Extract Derivations class to track proof metadata
